Arreglado para que se muestre email de admin en home, agregado botón de eliminar comentario(sin función).

+++ Controller/MarksController.php
@@ -30,7 +30,7 @@
         //INSERTA UNA NUEVA MARCA
         function InsertMark(){
             $logeado = $this->loginController->CheckLoggedIn();
-            if($logeado){
+            if($logeado && $_SESSION['ADMIN'] == 1){
                 if (isset($_POST['input_mark']) && isset($_POST['input_category'])) {
                     $mark = $_POST['input_mark'];
                     $category = $_POST['input_category'];
@@ -44,7 +44,7 @@
         //ELIMINA UNA MARCA POR ID
         function DeleteMark($params = null){
             $logeado = $this->loginController->CheckLoggedIn();
-            if($logeado){
+            if($logeado && $_SESSION['ADMIN'] == 1){
                 $mark_id = $params[':ID'];
                 $this->marksModel->DeleteMark($mark_id);
                 $this->productsView->ShowLocation('admin');
@@ -55,7 +55,7 @@
         //LLAMA LA VISTA PARA EDITAR UNA MARCA POR ID
         function EditMark($params = null){
             $logeado = $this->loginController->CheckLoggedIn();
-            if($logeado){
+            if($logeado && $_SESSION['ADMIN'] == 1){
                 $mark_id = $params[':ID'];
                 $mark = $this->marksModel->GetMarkById($mark_id);
                 $this->marksView->ShowEditMark($mark);
@@ -66,7 +66,7 @@
         //LLAMA A ACTUALIZAR UNA MARCA
         function UpdateMark($params = null){
             $logeado = $this->loginController->CheckLoggedIn();
-            if($logeado){
+            if($logeado && $_SESSION['ADMIN'] == 1){
                 $mark_id = $params[':ID'];
                 if (isset($_POST['edit_mark']) && isset($_POST['edit_category'])) {
                     $mark = $_POST['edit_mark'];
