+++ Controller/UserController.php
@@ -18,7 +18,7 @@
         //VISTA PARA EL ADMINISTRADOR DE LOS USUARIOS
         function ShowUsers(){
             $logeado = $this->loginControl->checkLoggedIn();
-            if($logeado){
+            if($logeado && $_SESSION['ADMIN'] == 1){
                 $users = $this->model->GetUsers();
                 $this->view->ShowUsers($users);
             }else{
@@ -28,7 +28,7 @@
         //LLAMA A LA VISTA PARA EDITAR UN USUARIO (SOLO PARA ADMINISTRADORES)
         function EditUser($params = null){
             $logeado = $this->loginControl->checkLoggedIn();
-            if($logeado){
+            if($logeado && $_SESSION['ADMIN'] == 1){
                 $id = $params[':ID'];
                 $user = $this->model->GetUserById($id);
                 $this->view->ShowEdit($user);
@@ -39,7 +39,7 @@
         //ACTUALIZA LOS DATOS DE UN USUARIO (SOLO PARA ADMINISTRADORES)
         function UpdateUser($params = null){
             $logeado = $this->loginControl->checkLoggedIn();
-            if($logeado){
+            if($logeado && $_SESSION['ADMIN'] == 1){
                 $id = $params[':ID'];
                 if(isset($_POST['selectAdmin'])){
                     $admin = $_POST['selectAdmin'];
@@ -53,7 +53,7 @@
         //ELIMINA UN USUARIO (SOLO PARA ADMINISTRADORES)
         function DeleteUser($params = null){
             $logeado = $this->loginControl->CheckLoggedIn();
-            if($logeado){
+            if($logeado && $_SESSION['ADMIN'] == 1){
                 $id = $params[':ID'];
                 $this->model->DeleteUser($id);
                 header("Location: ".BASE_URL.adminUsers);
